RELEASE 1.0.17+: Rename GameObjects namespace to Objects. Move Core/HarmonyPatches namespace to Harmony namespace.
